refactor: unify dummy mappings functions of export and start
